+++ src/QueryMap/Component/Map/QueryMapInterface.php
@@ -3,26 +3,7 @@ namespace QueryMap\Component\Map;
 
 interface QueryMapInterface
 {
-    /**
-     * Add a set of filters to the QueryMap buffer
-     *
-     * @param  array  $filters    A key-value array with filters. When a join filter
-     *                            is given its value can be another array of filters
-     * @return QueryMapInterface
-     */
-    public function add(array $filters);
-
-    /**
-     * Applies the added set of filters to the QueryMap and returns a query object
-     */
-    public function make();
-
-    /**
-     * Recreates the (engine-specific) query object of the QueryMap (but not the buffer)
-     *
-     * @return QueryMap
-     */
-    public function fresh();
+    public function query(array $filters, $reuse = false);
 
     /**
      * Returns the alias of the table
@@ -36,17 +17,6 @@ interface QueryMapInterface
      */
     public function getMappingHelper();
 
-    /**
-     * Returns the filter buffer (i.e. filters about to be used)
-     * @return array
-     */
-    public function getBuffer();
-
-    /**
-     * Resets the filter buffer so that an instance can be reused
-     * @return QueryMapInterface
-     */
-    public function resetBuffer();
 
     /**
      * Loads the filters from class annotations