+++ tests/Doctrine/Test/CreatureTest.php
@@ -21,25 +21,23 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'age__gte' => 16,
-                'age__lte' => 30,
-                'species__neq' => 'Human',
-                'status' => 'alive',
-                'netWorth__gt' => 10000,
-                'net_worth__lt' => 30000 // test: alias from column field
-            )
-        )->make();
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
+        $query = $creatureQueryMap->query([
+            'age__gte' => 16,
+            'age__lte' => 30,
+            'species__neq' => 'Human',
+            'status' => 'alive',
+            'netWorth__gt' => 10000,
+            'net_worth__lt' => 30000 // test: alias from column field
+        ]);
 
-        $this->assertStringMatchesFormat('%S(%s.age >= 16)%S', $resultingSql, '!gte');
-        $this->assertStringMatchesFormat('%S(%s.age <= 30)%S', $resultingSql, '!lte');
-        $this->assertStringMatchesFormat("%S(%s.status = 'alive')%S", $resultingSql, '!eq');
-        $this->assertStringMatchesFormat("%S(%s.species <> 'Human')%S", $resultingSql, '!neq');
-        $this->assertStringMatchesFormat('%S(%s.net_worth > 10000)%S', $resultingSql, '!gt');
-        $this->assertStringMatchesFormat('%S(%s.net_worth < 30000)%S', $resultingSql, '!lt');
+        $dql = $query->getDQL();
+        $this->assertContains('(cr.age >= 16)', $dql, '!gte');
+        $this->assertContains('(cr.age <= 30)', $dql, '!lte');
+        $this->assertContains("(cr.status = 'alive')", $dql, '!eq');
+        $this->assertContains("(cr.species <> 'Human')", $dql, '!neq');
+        $this->assertContains('(cr.netWorth > 10000)', $dql, '!gt');
+        $this->assertContains('(cr.netWorth < 30000)', $dql, '!lt');
     }
 
     public function testJoinOperatorsCanWorkAsFiltersWithoutJoiningAndTestOtherOperators()
@@ -51,28 +49,25 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
         // alias since an entity manager has to be specified anyway
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'status__in' => array('alive', 'zombie'),
-                'faction_id' => null, // => IS NULL
-                'species__eq' => 'Indogene',
-                'race' => 1,
-                'arrivalDate__neq' => null // => IS NOT NULL
-            )
-        );
+        $query = $creatureQueryMap->query([
+            'status__in' => array('alive', 'zombie'),
+            'faction_id' => null, // => IS NULL
+            'species__eq' => 'Indogene',
+            'race' => 1,
+            'arrivalDate__neq' => null // => IS NOT NULL
+        ]);
+
         // test: calling add multiple times appends to filters
-        $creatureQueryMap->add(array('name__like' => '%Yewkhaji'));
-        $creatureQueryMap->make();
+        $creatureQueryMap->query(['name__like' => '%Yewkhaji'], true);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-
-        $this->assertNotContains('[IGNORED_NULL]', $resultingSql, '!null');
-        $this->assertStringMatchesFormat("%S(%s.status IN ('alive', 'zombie'))%S", $resultingSql, '!in');
-        $this->assertStringMatchesFormat('%S(%s.faction_id IS NULL)%S', $resultingSql, '!null');
-        $this->assertStringMatchesFormat("%S(%s.species = 'Indogene')%S", $resultingSql, '!eq');
-        $this->assertStringMatchesFormat('%S(%s.race_id = 1)%S', $resultingSql, '!eq>join_col');
-        $this->assertStringMatchesFormat('%S(%s.arrival_date IS NOT NULL)%S', $resultingSql, '!notnull');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE '%Yewkhaji')%S", $resultingSql, '!notnull');
+        $dql = $query->getDQL();
+        $this->assertNotContains('[IGNORED_NULL]', $dql, '!null');
+        $this->assertContains("(cr.status IN ('alive', 'zombie'))", $dql, '!in');
+        $this->assertContains('(cr.faction IS NULL)', $dql, '!null');
+        $this->assertContains("(cr.species = 'Indogene')", $dql, '!eq');
+        $this->assertContains('(cr.race = 1)', $dql, '!eq>join_col');
+        $this->assertContains('(cr.arrivalDate IS NOT NULL)', $dql, '!notnull');
+        $this->assertContains("(cr.name LIKE '%Yewkhaji')", $dql, '!notnull');
     }
 
     public function testSeveralSimpleJoinCases()
@@ -81,58 +76,48 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
         // create a query to get all creatures whose race lives in a planet with breathable air
         // test: mix between left and inner joins
-        $creatureQueryMap->add(
-            array(
-                'race__ijo' => array(
-                    'name__contains' => 'Omec',
-                    'homeland__ljo' => array(  // test: the alias for homePlanet
-                        'atmosphereType' => 'N2O2CO2H2O'
-                    )
-                )
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'race__ijo' => [ // the alias is expected to be the first two characters of the property: 'ra'. since no explicit one was given
+                'name__contains' => 'Omec',
+                'homeland__ljo' => ['atmosphereType' => 'N2O2CO2H2O'] // test: the alias for homePlanet => alias will be 'ho'
+            ]
+        ]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-
-        $this->assertStringMatchesFormat('%SINNER JOIN races %s ON %s.race_id = %s.id%S', $resultingSql, '!ijo');
-        $this->assertStringMatchesFormat('%SLEFT JOIN planets %s ON %s.home_planet = %s.id%S', $resultingSql, '!ljo');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE '%Omec%')%S", $resultingSql, '!any');
-        $this->assertStringMatchesFormat("%S(%s.atmosphere_type = 'N2O2CO2H2O')%S", $resultingSql, '!ljo>eq');
+        $dql = $query->getDQL();
+        $this->assertContains('INNER JOIN cr.race ra', $dql, '!ijo');
+        $this->assertContains('LEFT JOIN ra.homePlanet ', $dql, '!ljo');
+        $this->assertContains("(ra.name LIKE '%Omec%')", $dql, '!any');
+        $this->assertContains("(ho.atmosphereType = 'N2O2CO2H2O')", $dql, '!ljo>eq');
     }
 
     public function testMultipleJoinsAndAliasNamingConflictDoesNotHappen()
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'faction__ljo' => array(
-                    //the reason to declare id columns as @Filter is to allow
-                    //joining with a given id.
-                    'id' => 3
-                )
-            )
-        )->make();
+        $creatureQueryMap->query([
+            'faction__ljo' => [
+                //the reason to declare id columns as @Filter is to allow
+                //joining with a given id.
+                'id' => 3
+            ]
+        ], $creatureQueryMap::REUSE_FILTERS);
 
         // since Doctrine picks creates its own aliases and we don't need to
         // specify an alias by which to join in our entities, there is no
         // chance of naming collision here; the joins will be performed separately
-        $creatureQueryMap->add(
-            array(
-                'race__ijo' => array(
-                    'leadingFaction__ijo' => array(
-                        'capital' => 'Bucharest'
-                    )
-                )
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'race__ijo' => [
+                'leadingFaction__ijo' => [
+                    'capital' => 'Bucharest'
+                ]
+            ]
+        ]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-
-        $this->assertStringMatchesFormat('%SLEFT JOIN factions %s ON %s.faction_id = %s.id%S', $resultingSql, '!ljo');
-        $this->assertStringMatchesFormat('%SINNER JOIN races %s ON %s.race_id = %s.id%S', $resultingSql, '!ijo_1');
-        $this->assertStringMatchesFormat('%SINNER JOIN factions %s ON %s.faction_id = %s.id%S', $resultingSql, '!ijo_2');
-        $this->assertStringMatchesFormat('%S(%s.id = 3)%S', $resultingSql, '!ljo>eq');
+        $dql = $query->getDQL();
+        $this->assertContains('LEFT JOIN cr.faction fa', $dql, '!ljo');
+        $this->assertContains('INNER JOIN cr.race ra', $dql, '!ijo_1');
+        $this->assertContains('INNER JOIN ra.leadingFaction le', $dql, '!ijo_2');
+        $this->assertContains('(fa.id = 3)', $dql, '!ljo>eq');
     }
 
     public function testMultipleJoinsCircularDependency()
@@ -141,33 +126,28 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
 
         //should work as long as there is no aliasing conflict
-        $creatureQueryMap->add(
-            array(
-                'name__like' => 'Gigel%',
-                'faction__ljo' => array(
-                    //the reason to declare id columns as @Filter is to allow
-                    //joining with a given id.
-                    'guns__gte' => 500,
-                    'capital__ijo' => array(
-                        'leading_race__ijo' => array(
-                            'home_planet__ijo' => array(
-                                'name__contains' => 'Chiajna'
-                            )
-                        )
-                    )
-                )
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'name__like' => 'Gigel%',
+            'faction__ljo' => [
+                'guns__gte' => 500,
+                'capital__ijo' => [
+                    'leading_race__ijo' => [
+                        'home_planet__ijo' => [
+                            'name__contains' => 'Chiajna'
+                        ]
+                    ]
+                ]
+            ]
+        ]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-
-        $this->assertStringMatchesFormat('%SLEFT JOIN factions %s ON %s.faction_id = %s.id%S', $resultingSql, '!ljo');
-        $this->assertStringMatchesFormat('%SINNER JOIN planets %s ON %s.capital = %s.id%S', $resultingSql, '!ijo_1');
-        $this->assertStringMatchesFormat('%SINNER JOIN races %s ON %s.leading_race = %s.id%S', $resultingSql, '!ijo_2');
-        $this->assertStringMatchesFormat('%SINNER JOIN planets %s ON %s.home_planet = %s.id%S', $resultingSql, '!ijo_3');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE 'Gigel%')%S", $resultingSql, '!like');
-        $this->assertStringMatchesFormat('%S(%s.gun_count >= 500)%S', $resultingSql, '!ljo>gte');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE '%Chiajna%')%S", $resultingSql, '!ijo>like');
+        $dql = $query->getDQL();
+        $this->assertContains('LEFT JOIN cr.faction fa', $dql, '!ljo');
+        $this->assertContains('INNER JOIN fa.capital ca', $dql, '!ijo_1');
+        $this->assertContains('INNER JOIN ca.leadingRace le', $dql, '!ijo_2');
+        $this->assertContains('INNER JOIN le.homePlanet ho', $dql, '!ijo_3');
+        $this->assertContains("(cr.name LIKE 'Gigel%')", $dql, '!like');
+        $this->assertContains('(fa.guns >= 500)', $dql, '!ljo>gte');
+        $this->assertContains("(ho.name LIKE '%Chiajna%')", $dql, '!ijo>like');
     }
 
     /**
@@ -178,7 +158,7 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(array('arivalDate_gte' => '12-10-2015'))->make();
+        $creatureQueryMap->query(array('arivalDate_gte' => '12-10-2015'));
     }
 
     /**
@@ -189,7 +169,7 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(array('arrivalDate__geq' => '12-10-2015'))->make();
+        $creatureQueryMap->query(array('arrivalDate__geq' => '12-10-2015'));
     }
 
     public function testNoErrorWhenJoinFilterHasNoQueryMap()
@@ -198,38 +178,40 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
         //inner
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(array('secondFaction__ijo' => array()))->make();
+        $query = $creatureQueryMap->query(['secondFaction__ijo' => []]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-        $this->assertStringMatchesFormat('%SINNER JOIN factions %s ON %s.second_faction_id = %s.id%S', $resultingSql, '!ijo');
+        $dql = $query->getDQL();
+        $this->assertContains('INNER JOIN cr.secondFaction se', $dql, '!ijo');
 
         //left
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(array('secondFaction__ljo' => array()))->make();
+        $query = $creatureQueryMap->query(['secondFaction__ljo' => []]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-        $this->assertStringMatchesFormat('%SLEFT JOIN factions %s ON %s.second_faction_id = %s.id%S', $resultingSql, '!ljo');
+        $dql = $query->getDQL();
+        $this->assertContains('LEFT JOIN cr.secondFaction se', $dql, '!ijo');
     }
 
     public function testMethodFiltersReallyAddSomethingToTheQuery()
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                // test: custom 'between' with alias and explicit suffix (optional, just for fun)
-                'creatureArrivedBetween__method' => array('12-10-2014', '12-10-2015'),
-                'faction__ijo' => array(
-                    'creatureShareGreaterThan' => 10
-                    //'creatureShare__gt' > 10  // should it be possible to combine method operator with something else?
-                )
-            )
-        )->make();
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
+        $query = $creatureQueryMap->query([
+            // test: custom 'between' with alias and explicit suffix (optional, just for fun)
+            'creatureArrivedBetween__method' => ['12-10-2014', '12-10-2015'],
+            'faction__ijo' => [
+                'creatureShareGreaterThan' => 10
+                //'creatureShare__gt' > 10  // should it be possible to combine method operator with something else?
+            ]
+        ]);
 
-        $this->assertStringMatchesFormat("%S(%s.arrival_date BETWEEN '12-10-2014' AND '12-10-2015')%S", $resultingSql, '!method');
-        $this->assertStringMatchesFormat('%S(((%s.net_worth * 100) / %s.net_worth) > 10)%S', $resultingSql, '!method2');
+        $dql = $query->getDQL();
+        $this->assertContains("(cr.arrivalDate BETWEEN :arrivalDateStart AND :arrivalDateEnd)", $dql, '!method');
+        $this->assertContains('(((cr.netWorth * 100) / fa.netWorth) > :percent)', $dql, '!method2');
+
+        $this->assertEquals('12-10-2014', $query->getParameter('arrivalDateStart')->getValue());
+        $this->assertEquals('12-10-2015', $query->getParameter('arrivalDateEnd')->getValue());
+        $this->assertEquals('10', $query->getParameter('percent')->getValue());
     }
 
     /**
@@ -240,7 +222,7 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(array('badMethodFilter' => 1337))->make();
+        $creatureQueryMap->query(['badMethodFilter' => 1337]);
     }
 
     /**
@@ -251,23 +233,21 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'creature_parent__ijo' => array(
-                    'name__like' => '%Wurtt',
-                    'creature_parent__ljo' => array(
-                        'name__contains' => 'Grandpa'
-                    )
-                )
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'creature_parent__ijo' => [
+                'name__like' => '%Wurtt',
+                'creature_parent__ljo' => [
+                    'name__contains' => 'Grandpa'
+                ]
+            ]
+        ]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
 
-        $this->assertStringMatchesFormat('%SINNER JOIN creatures %s ON %s.creature_parent = %s.id%S', $resultingSql, '!collision_1');
-        $this->assertStringMatchesFormat('%SLEFT JOIN creatures %s ON %s.creature_parent = %s.id%S', $resultingSql, '!collision_2');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE '%Wurtt')%S", $resultingSql, '!collision_3');
-        $this->assertStringMatchesFormat("%S(%s.name LIKE '%Grandpa%')%S", $resultingSql, '!collision_4');
+        $dql = $query->getDQL();
+        $this->assertContains('INNER JOIN cr.aCreatureParent aC', $dql, '!collision_1');
+        $this->assertContains('LEFT JOIN aC.aCreatureParent aC2', $dql, '!collision_2');
+        $this->assertContains("(aC.name LIKE '%Wurtt')", $dql, '!collision_3');
+        $this->assertContains("(aC2.name LIKE '%Grandpa%')", $dql, '!collision_4');
     }
 
     /**
@@ -281,11 +261,7 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
 
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'race' => $race
-            )
-        )->make();
+        $creatureQueryMap->query(['race' => $race]);
     }
 
     /**
@@ -295,103 +271,76 @@ class CreatureTest extends \PHPUnit_Framework_TestCase
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'cash__between' => array(10000, 20000),
-                'arrivalDate__between' => array('24-12-2015', '31-12-2015')
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'cash__between' => [10000, 20000],
+            'arrivalDate__between' => ['24-12-2015', '31-12-2015']
+        ]);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
 
-        $this->assertStringMatchesFormat('%Snet_worth BETWEEN 10000 AND 20000%S', $resultingSql, '!between1');
-        $this->assertStringMatchesFormat("%Sarrival_date BETWEEN '24-12-2015' AND '31-12-2015'%S", $resultingSql, '!between2');
+        $dql = $query->getDQL();
+        $this->assertContains('cr.netWorth BETWEEN 10000 AND 20000', $dql, '!between1');
+        $this->assertContains("cr.arrivalDate BETWEEN '24-12-2015' AND '31-12-2015'", $dql, '!between2');
     }
 
     public function testAliasForJoin()
     {
         /** @var \QueryMap\Component\Map\QueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
-        $creatureQueryMap->add(
-            array(
-                'faction__ijo__ftn' => array(),
-                'race' => 5     // bonus test: join filter without join
-            )
-        )->make();
+        $query = $creatureQueryMap->query([
+            'faction__ijo__ftn' => [],
+            'race' => 5     // bonus test: join filter without join
+        ]);
 
-        /** @var \Doctrine\ORM\QueryBuilder $queryBuilder */
-        $queryBuilder = $creatureQueryMap->getQuery();
-        $resultingDql = $queryBuilder->getDQL();
-
-        $this->assertContains('INNER JOIN cr.faction ftn', $resultingDql);
-        $this->assertContains('(cr.race = 5)', $resultingDql);
+        $dql = $query->getDQL();
+        $this->assertContains('INNER JOIN cr.faction ftn', $dql);
+        $this->assertContains('(cr.race = 5)', $dql);
     }
 
     public function testCommonMappingHelper()
     {
-        $filtersRaw = array(
+        $filtersRaw = [
             'arrived-before' => '24-12-2020',
             'cash-at-least' => '10000',
             'has-faction-share-above' => '20%'
-        );
+        ];
 
-        $mapping = array(
-            'arrived-before' => array(
-                'preProcess' => array(CommonMappingHelper::TYPE_STRING, CommonMappingHelper::TRANSFORM_TRIM),
-                'validate' => array(CommonMappingHelper::VALID_DATETIME_STRING),
-                'process' => array(
-                    CommonMappingHelper::TRANSFORM_DATETIME => array(
+        $mapping = [
+            'arrived-before' => [
+                'preProcess' => [CommonMappingHelper::TYPE_STRING, CommonMappingHelper::TRANSFORM_TRIM],
+                'validate' => [CommonMappingHelper::VALID_DATETIME_STRING],
+                'process' => [
+                    CommonMappingHelper::TRANSFORM_DATETIME => [
                         CommonMappingHelper::SET_TIME_MAX
-                    )
-                ),
-                'key' => array('arrivalDate__lte' => null)
-            ),
-            'cash-at-least' => array(
-                'preProcess' => array(CommonMappingHelper::TYPE_INT),
-                'validate' => array(CommonMappingHelper::VALID_NOT_EMPTY),
-                'key' => array('cash__gte' => null)
-            ),
-            'has-faction-share-above' => array(
-                'preProcess' => array(CommonMappingHelper::TYPE_INT),
-                'validate' => array(CommonMappingHelper::VALID_NOT_EMPTY),
-                'key' => array(
-                    'faction__ijo' => array(
+                    ]
+                ],
+                'key' => ['arrivalDate__lte' => null]
+            ],
+            'cash-at-least' => [
+                'preProcess' => [CommonMappingHelper::TYPE_INT],
+                'validate' => [CommonMappingHelper::VALID_NOT_EMPTY],
+                'key' => ['cash__gte' => null]
+            ],
+            'has-faction-share-above' => [
+                'preProcess' => [CommonMappingHelper::TYPE_INT],
+                'validate' => [CommonMappingHelper::VALID_NOT_EMPTY],
+                'key' => [
+                    'faction__ijo' => [
                         'creatureShareGreaterThan' => null
-                    )
-                )
-            )
-        );
+                    ]
+                ]
+            ]
+        ];
 
         /** @var \QueryMap\Contrib\Map\CommonQueryMap $creatureQueryMap */
         $creatureQueryMap = $this->service->create(Creature::class, 'cr');
 
         $filters = $creatureQueryMap->transform($filtersRaw, $mapping);
-        $creatureQueryMap->add($filters)->make();
+        $query = $creatureQueryMap->query($filters);
 
-        $resultingSql = static::sqlNormalize($creatureQueryMap->getQuerySql());
-
-        $this->assertStringMatchesFormat("%Sarrival_date <= '2020-12-24 23:59:59%S", $resultingSql, '!arrivedBefore');
-        $this->assertStringMatchesFormat('%Snet_worth >= 10000%S', $resultingSql, '!cashAtLeast');
-        $this->assertStringMatchesFormat('%S(((%s.net_worth * 100) / %s.net_worth) > 20)%S', $resultingSql, '!shareGt');
-    }
-
-    /**
-     * Defancify query
-     * @param $sql
-     * @return mixed
-     */
-    static protected function sqlNormalize($sql)
-    {
-        $sql = str_replace("\n", '', $sql);
-
-        while ($sql !== ($new = str_replace('  ', ' ', $sql))) {
-            $sql = $new;
-        }
-
-        $sql = str_replace('( ', '(', $sql);
-        $sql = str_replace(' )', ')', $sql);
-        $sql = str_replace('`', '', $sql);
-
-        return $sql;
+        $dql = $query->getDQL();
+        $this->assertContains("cr.arrivalDate <= '2020-12-24 23:59:59", $dql, '!arrivedBefore');
+        $this->assertContains('cr.netWorth >= 10000', $dql, '!cashAtLeast');
+        $this->assertContains('(((cr.netWorth * 100) / fa.netWorth) > :percent)', $dql, '!shareGt');
+        $this->assertEquals('20', $query->getParameter('percent')->getValue());
     }
 }